working on encoder

+++ PongGame/PongGame/p1000_driver.h
@@ -13,7 +13,7 @@
 #include "ATmega162_driver.h"
 #include "ADC_driver.h"
 
-typedef enum pos_t {LEFT, RIGHT, UP, DOWN, NEUTRAL} pos_t;void joystick_analog_position(float* x_per, float* y_per, uint8_t* ADC_data, int* calibrated);void slider_position(float *left_per, float *right_per, uint8_t* ADC_data);pos_t pos_read(float* x_per, float* y_per);
+typedef enum pos_t {LEFT, RIGHT, UP, DOWN, NEUTRAL} pos_t;void joystick_analog_position(float* x_per, float* y_per, uint8_t* ADC_data, int* calibrated);void slider_position(int *left_per, int *right_per, uint8_t* ADC_data);pos_t pos_read(float* x_per, float* y_per);
 
 
 #endif /* P1000_DRIVER_H_ */
